Restructure model and add output logging.
The model logic was revisted and reworked in several ways, with the goal
of improving class encapsulation and making logging of model outputs
more acheivable. Several gaps in the model logic were also discovered
and fixed. Overall, changes included:

	- Addition of ModelLogger class to write model output to JSON
	- Moved patient resource allocation and treatment to Hospital
	- Combined returning of human resources with treatment
	- Treat patients with already assigned staff first
	- Added notification method for patient death in Hospital
	- Made available nurses and doctors in Hospital private members
	- Added shift management for doctors and nurses in Hospital
	- Added treatment status cache for patients in Hospital
	- Added fields for active nurses and doctors in Patient
	- Added logic for treatment completion and next treatment
	- Added getLength() to TreatmentPlan to return total cycles
	- Moved Stage to a static inner class of StageManager
	- Added cycle counting to Stage object to track duration of stage

ModelLogger class was implemented as a static class as only a single
instance of the logger should be needed. The class contains a private
internal class, PatientSerializer, that is resposible for returning JSON
objects for specific aspects of logging patient data such as initial
admission, cycle-to-cycle status, and patient death. The class has two
public member functions, newCycle() and logPatient(). The newCycle()
method should be invoked every time subsequent patient logging should be
attributed to a new cycle, and logPatient() should be called to log
patient data for a specific patient. The logPatient() function
automatically checks whether the patient was just admitted or if the
patient has died, and logs that data appropriately using the
PatientSerializer.

Resource allocation, patient treatment, and returning of human resources
was all moved to a single method in Hospital, attemptTreatments(). This
was done to improve encapsulation for the Hospital object, which
previously had public members for available nurses and doctors. These
properties are now private (m_nurses and m_doctors), making the Hospital
class the only class that is able to directly modify human and material
resources. Once resources have been allocated and all patients have been
treated or not treated, human resources are returned to the hospital
pool based on whether they are needed again by any given patient in the
next cycle. This decision is now made directly in the Hospital class
instead of the TreatmentPlan. The related methods
freeNursesAfterTreatment() and freeDcotorsAfterTreatment() were
consequently removed. The TreatmentPlan class now only provides
information about what resources are required for the current or next
cycle. The hospital may use this information to determine whether
doctors or nurses should be left with the patient or returned to the
resource pool. Properties for active nurses and doctors that are with a
patient were added to the Patient class. These properties are updated
when a nurse or doctor is allocated to a patient or returned to the
hospital.

The treatment priority for patients in Hospital was also updated to
treat those patients with already assigned nurses and doctors first.
This occurs when the hospital leaves nurses or doctors with a patient if
they are needed for consecutive cycles. These patients are treated first
as they already have staff assigned, and so that if they are unable to
be treated due to missing material resources, their nurses and/or
doctors are able to be reassigned to another patient.

Shift management for doctors and nurses was also added to the Hospital
class through the methods setNursesOnShift() and setDoctorsOnShift().
This allows for the number of doctors and nurses that are on shift at
any given time to be modified. If the total number of staff on shift
increases, the additional members are immediately available. If the
total number decreases, doctors or nurses are releived from their shift
as soon as they are done visiting their current patient. This logic is
implemented in the attemptTreatments() method when human resources are
being returned to the hospital after treating patients.

Moved treatment to Hospital object within attemptTreatments(), to
improve encapsulation; the hospital should be attempting to treat
patients and managing its own human and material resources. The number
of free doctors and nurses was consequently made read only through
accessor methods instead of being public.

Logic for returning hospital human resources was also combined with the
attemptTreatments() method to support encapsulation.

Added shift support for doctors and nurses, so that the number of
doctors and nurses on shift at any given time can be modified.
Additional doctors and nurses are added immediately to the number of
available doctors and nurses, while their removal occurs as they become
available after treating their current patients.

The method notifyPatientDeath() was added to Hospital so that the
hospital has the opportunity to reclaim any human resources that were
assigned to a patient that just died.

The private member m_patientTreatmentStatus was added to Hospital as a
HashMap to track which patients were treated and which were not, for any
given cycle. The HashMap persists between cycles and is simply updated
during attemptTreatments(). The method didTreat() was added to suplement
this member and return the most recent treatment status for a specific
patient.

The setup.xlsx spreadsheet was updated to include an addition column,
"Next Treatment", that specifies which treatment plan should be used
next for a patient once they have completed their current treatment
plan. Note that the next treatment plan should still have the same match
criteria in order for the patient to persist in that treatment in
subsequent cycles. The appropriate logic to check for whether the
patient has completed their current treatment plan was added to
StageManager.

The Stage class was moved to be a static inner class of StageManager.
This was done in order to add a private member m_cycle within Stage that
is updated each time StageManager.getStage() is called in order to track
the length of time (in cycles) that a patient has been in a particular
stage. By being an inner class of StageManager, only the StageManager
class is able to access and modify m_cycle. This property is used by
StageManager to determine whether a patient has completed a treatment
plan based on the length or total cycles specified for each treatment
plan.

Also updated .gitignore to ignore excel auxillary files for the
setup.xlsx spreadsheet (~$setup.xlsx).