Initial commit for a separate model visualizer application.
The model visualizer application was added as a separate package that
allows users to parse through data output files from the model and
generate plots quickly and easily.

The main class file for the visualizer is the App class, which currently
just reads in the first file in the output folder for analysis. The
class then creates an instance of the ModelVisualizer class to parse
through and visualize the data from the output file. The ModelVisulizer
class utilizes XChart and MigLayout as dependencies to plot the data and
easily manage visual layouts, respectively.

The App class from the MMHModel package was also updated to include a
static method that simply returns the output folder. This method is used
by the App class from the ModelVisualizer package to list the contents
of the output folder and choose a file for analysis.