Update condition logic for initial POM and hospital inventory.
The initial probability of mortality for any given condition assigned to
a new patient was orginally randomly chosen between 0 and 1 using a
uniform probability distribution. This did not accurately reflect the
distribution of severities that are typically seen among patients at
admission. This implementation was consequently updated to randomly
choose a probability of mortality from a custom gaussian probabilty
distribution, where the mean and standard deviation for the distribution
is unique for each condition and is defined in the first tab of the
setup.xlsx file. This initial probability of mortality is still
generated within the createPatient() function in the Shiva class. Since
a gaussian distribution is not bounded in domain, the helper function
getProbabilityOfMortality() was added to draw a random number until it
falls within the range of 0 and 1 exclusively.

The enumeration entry for misoprostol was misspelled and corrected. A
corresponding hospital inventory stock for misoprostol was found to be
missing and was added.

The RUPTURED_UTERUS enumeration entry for Condition.Type was renamed to
RU.

The setup.xlsx file was updated to include a mean and standard deviation
for the initial probability of mortility of each condition. The
condition incidences and list of trajectories was also updated based on
information collected from MMH and provided by Benoit.