Implement patient discharge and change treatment logic.
A public static final Stage object (Stage.Complete) was created to
provide a single stage object that is returned by the StageManager when
a patient has completed all of their stages, signified by a treatment ID
of 0 in the trajectories list of the setup.xlsx file.

Simulation logic was updated to check if a patient is in the Complete
stage, and discharges the patient from the hospital if so by removing
them from the current patient list.

The treatment logic in TreatmentPlan was updated so that the new
probability of mortality for a given condition after treatment follows a
geometric decay that is independent of the current treatment cycle.