Stock and probability of mortality logic changes.
Additional items added to the hospital stock that were missing, and the
disposable resources enumeration and stocking of resources was reordered
to alphabetical ordering.

Probability of mortality logic was altered for determining whether a
patient dies each patient. While the probability of mortality reflects
the overall chance for a patient to die during their hospital stay,
using this probability to evaluate whether a patient survives any given
cycle results in a signficantly higher cumulative probability that the
patient dies over the coures of all hospital cycles. This can be
represented by a binomial distribution, where the probability of
mortality is 1 - Pr(0;n,p) or the inverse probability of not dying over
any of the cycles, where p is the probability of dying at any given
cycle, and n is the number of remaining cycles during the patient's
total stay. The simulation logic was thus altered to calculate p from
the cumulative probability at each cycle. This calculated probability is
then used to draw a random variate and determine whether the patient
survives the current cycle.

A function was added to the StageManager to determine the total number
of remaining cycles that a patient has based on their current stage and
set of conditions.